Gave Iron Ingot an actual collider. Fixed a bug where items could be dropped and the smelter could be placed in the air. Fixed a minor bug with Inventory description text.

+++ Assets/Code/UI/InventoryActions/InventoryDrop.cs
@@ -7,13 +7,16 @@ public class InventoryDrop : InventoryAction
 {
     public override void InventoryAct(Inventory inventory){
         World world = inventory.GetWorld();
-        Vector3 modifiedPosition = world.GetPlayer().transform.position;
-        ItemSlot item = inventory.GetSelected();
-        modifiedPosition.y -= 0.5f;
-        Debug.Log("Inventory Drop Called");
-        BasicItem newItemObject = Instantiate(world.GetPrefab(item.GetName()), modifiedPosition, Quaternion.identity).GetComponent<BasicItem>();
-        newItemObject.SetInventoryRef(inventory);
-        newItemObject.SetCount(item.GetCount());
-        inventory.DeleteItem();
+        if(world.GetPlayer().IsGrounded()){
+            Vector3 modifiedPosition = world.GetPlayer().transform.position;
+            ItemSlot item = inventory.GetSelected();
+            modifiedPosition.y -= 0.5f;
+            Debug.Log("Inventory Drop Called");
+            BasicItem newItemObject = Instantiate(world.GetPrefab(item.GetName()), modifiedPosition, Quaternion.identity).GetComponent<BasicItem>();
+            newItemObject.SetInventoryRef(inventory);
+            newItemObject.SetCount(item.GetCount());
+            inventory.DeleteItem();
+        } else inventory.DisplayError("You must be grounded to drop an item!");
+        
     }
 }
